@@ -1,8 +1,8 @@
 [package]
 name = "rainfrog"
-version = "0.3.2"
-edition = "2021"
-rust-version = "1.80"
+version = "0.3.5"
+edition = "2024"
+rust-version = "1.88"
 description = "a database management tui"
 repository = "https://github.com/achristmascarl/rainfrog"
 authors = ["achristmascarl <carl@rainfrog.dev>"]
@@ -80,7 +80,7 @@ tracing-subscriber = { version = "0.3.17", features = ["env-filter", "serde"] }
 chrono = { version = "0.4", default-features = false }
 indexmap = "2.2.6"
 tui-textarea = { version = "0.7.0", features = ["search"] }
-sqlparser = "0.56.0"
+sqlparser = "0.58.0"
 arboard = { version = "3.4.1", optional = true, features = [
   "wayland-data-control",
 ] }
