@@ -11,14 +11,16 @@ use color_eyre::eyre::{self, Result};
 use futures::stream::StreamExt;
 use sqlparser::ast::Statement;
 use sqlx::{
-  postgres::{PgConnectOptions, PgPoolOptions, Postgres},
-  types::Uuid,
   Column, Either, Row, ValueRef,
+  pool::PoolConnection,
+  postgres::{PgConnectOptions, PgConnection, PgPoolOptions, Postgres},
+  types::Uuid,
 };
+use tokio::sync::Mutex;
 use tokio::task::JoinHandle;
 
 use super::{
-  vec_to_string, Database, DbTaskResult, Driver, Header, Headers, QueryResultsWithMetadata, QueryTask, Rows, Value,
+  Database, DbTaskResult, Driver, Header, Headers, QueryResultsWithMetadata, QueryTask, Rows, Value, vec_to_string,
 };
 
 type PostgresTransaction<'a> = sqlx::Transaction<'a, Postgres>;
@@ -33,6 +35,8 @@ enum PostgresTask<'a> {
 pub struct PostgresDriver<'a> {
   pool: Option<Arc<sqlx::Pool<Postgres>>>,
   task: Option<PostgresTask<'a>>,
+  querying_conn: Option<Arc<Mutex<PoolConnection<Postgres>>>>,
+  querying_pid: Option<String>,
 }
 
 #[async_trait(?Send)]
@@ -46,11 +50,23 @@ impl Database for PostgresDriver<'_> {
 
   // since it's possible for raw_sql to execute multiple queries in a single string,
   // we only execute the first one and then drop the rest.
-  fn start_query(&mut self, query: String) -> Result<()> {
-    let (first_query, statement_type) = super::get_first_query(query, Driver::Postgres)?;
+  async fn start_query(&mut self, query: String, bypass_parser: bool) -> Result<()> {
+    let (first_query, statement_type) = match bypass_parser {
+      true => (query, None),
+      false => {
+        let (first, stmt) = super::get_first_query(query, Driver::Postgres)?;
+        (first, Some(stmt))
+      },
+    };
     let pool = self.pool.clone().unwrap();
+    self.querying_conn = Some(Arc::new(Mutex::new(pool.acquire().await?)));
+    let conn = self.querying_conn.clone().unwrap();
+    let conn_for_task = conn.clone();
+    let pid = sqlx::raw_sql("SELECT pg_backend_pid()").fetch_one(conn.lock().await.as_mut()).await?.get::<i32, _>(0);
+    log::info!("Starting query with PID {}", pid.clone());
+    self.querying_pid = Some(pid.to_string().clone());
     self.task = Some(PostgresTask::Query(tokio::spawn(async move {
-      let results = query_with_pool(pool, first_query.clone()).await;
+      let results = query_with_conn(conn_for_task.lock().await.as_mut(), first_query.clone()).await;
       match results {
         Ok(ref rows) => {
           log::info!("{:?} rows, {:?} affected", rows.rows.len(), rows.rows_affected);
@@ -64,16 +80,42 @@ impl Database for PostgresDriver<'_> {
     Ok(())
   }
 
-  fn abort_query(&mut self) -> Result<bool> {
-    if let Some(task) = self.task.take() {
-      match task {
-        PostgresTask::Query(handle) => handle.abort(),
-        PostgresTask::TxStart(handle) => handle.abort(),
-        _ => {},
-      };
-      Ok(true)
-    } else {
-      Ok(false)
+  async fn abort_query(&mut self) -> Result<bool> {
+    match self.task.take() {
+      Some(task) => {
+        match task {
+          PostgresTask::Query(handle) => handle.abort(),
+          PostgresTask::TxStart(handle) => handle.abort(),
+          _ => {},
+        };
+        if let Some(pid) = self.querying_pid.take() {
+          let result =
+            sqlx::raw_sql(&format!("SELECT pg_cancel_backend({pid})")).fetch_one(&*self.pool.clone().unwrap()).await;
+          let msg = match &result {
+            Ok(_) => "Successfully killed".to_string(),
+            Err(e) => format!("Failed to kill: {e:?}"),
+          };
+
+          let success = result.as_ref().is_ok_and(|r| r.try_get::<bool, _>(0).unwrap_or(false));
+          let status_string =
+            result.map_or("ERROR".to_string(), |r| r.try_get_unchecked::<String, _>(0).unwrap_or("ERROR".to_string()));
+
+          if !success {
+            log::warn!("Unexpected response when cancelling backend process with PID {pid}: {msg}");
+            log::warn!("Status: {status_string}");
+          } else {
+            log::info!("Tried to cancel backend process with PID {pid}: {msg}");
+            log::info!("Status: {status_string}");
+          }
+        }
+        self.querying_conn = None;
+        Ok(true)
+      },
+      _ => {
+        self.querying_conn = None;
+        self.querying_pid = None;
+        Ok(false)
+      },
     }
   }
 
@@ -85,6 +127,8 @@ impl Database for PostgresDriver<'_> {
           (DbTaskResult::Pending, Some(PostgresTask::Query(handle)))
         } else {
           let result = handle.await?;
+          self.querying_conn = None;
+          self.querying_pid = None;
           (DbTaskResult::Finished(result), None)
         }
       },
@@ -97,10 +141,19 @@ impl Database for PostgresDriver<'_> {
             Ok(rows) => rows.rows_affected,
             _ => None,
           };
-          (
-            DbTaskResult::ConfirmTx(rows_affected, result.statement_type.clone()),
-            Some(PostgresTask::TxPending(Box::new((tx, result)))),
-          )
+          match result {
+            // if tx failed to start, return the error immediately
+            QueryResultsWithMetadata { results: Err(e), statement_type } => {
+              log::error!("Transaction didn't start: {e:?}");
+              self.querying_conn = None;
+              self.querying_pid = None;
+              (DbTaskResult::Finished(QueryResultsWithMetadata { results: Err(e), statement_type }), None)
+            },
+            _ => (
+              DbTaskResult::ConfirmTx(rows_affected, result.statement_type.clone()),
+              Some(PostgresTask::TxPending(Box::new((tx, result)))),
+            ),
+          }
         }
       },
       Some(PostgresTask::TxPending(b)) => (DbTaskResult::Pending, Some(PostgresTask::TxPending(b))),
@@ -111,27 +164,30 @@ impl Database for PostgresDriver<'_> {
 
   async fn start_tx(&mut self, query: String) -> Result<()> {
     let (first_query, statement_type) = super::get_first_query(query, Driver::Postgres)?;
-    let tx = self.pool.as_mut().unwrap().begin().await?;
+    let mut tx = self.pool.clone().unwrap().begin().await?;
+    let pid = sqlx::raw_sql("SELECT pg_backend_pid()").fetch_one(&mut *tx).await?.get::<i32, _>(0);
+    log::info!("Starting transaction with PID {}", pid.clone());
+    self.querying_pid = Some(pid.to_string().clone());
     self.task = Some(PostgresTask::TxStart(tokio::spawn(async move {
       let (results, tx) = query_with_tx(tx, &first_query).await;
       match results {
         Ok(Either::Left(rows_affected)) => {
-          log::info!("{:?} rows affected", rows_affected);
+          log::info!("{rows_affected:?} rows affected");
           (
             QueryResultsWithMetadata {
               results: Ok(Rows { headers: vec![], rows: vec![], rows_affected: Some(rows_affected) }),
-              statement_type,
+              statement_type: Some(statement_type),
             },
             tx,
           )
         },
         Ok(Either::Right(rows)) => {
           log::info!("{:?} rows affected", rows.rows_affected);
-          (QueryResultsWithMetadata { results: Ok(rows), statement_type }, tx)
+          (QueryResultsWithMetadata { results: Ok(rows), statement_type: Some(statement_type) }, tx)
         },
         Err(e) => {
           log::error!("{e:?}");
-          (QueryResultsWithMetadata { results: Err(e), statement_type }, tx)
+          (QueryResultsWithMetadata { results: Err(e), statement_type: Some(statement_type) }, tx)
         },
       }
     })));
@@ -141,11 +197,16 @@ impl Database for PostgresDriver<'_> {
   async fn commit_tx(&mut self) -> Result<Option<QueryResultsWithMetadata>> {
     if !matches!(self.task, Some(PostgresTask::TxPending(_))) {
       Ok(None)
-    } else if let Some(PostgresTask::TxPending(b)) = self.task.take() {
-      b.0.commit().await?;
-      Ok(Some(b.1))
     } else {
-      Ok(None)
+      match self.task.take() {
+        Some(PostgresTask::TxPending(b)) => {
+          b.0.commit().await?;
+          self.querying_conn = None;
+          self.querying_pid = None;
+          Ok(Some(b.1))
+        },
+        _ => Ok(None),
+      }
     }
   }
 
@@ -153,6 +214,8 @@ impl Database for PostgresDriver<'_> {
     if let Some(PostgresTask::TxPending(b)) = self.task.take() {
       b.0.rollback().await?;
     }
+    self.querying_conn = None;
+    self.querying_pid = None;
     Ok(())
   }
 
@@ -171,35 +234,33 @@ impl Database for PostgresDriver<'_> {
   }
 
   fn preview_rows_query(&self, schema: &str, table: &str) -> String {
-    format!("select * from \"{}\".\"{}\" limit 100", schema, table)
+    format!("select * from \"{schema}\".\"{table}\" limit 100")
   }
 
   fn preview_columns_query(&self, schema: &str, table: &str) -> String {
     format!(
-      "select column_name, * from information_schema.columns where table_schema = '{}' and table_name = '{}'",
-      schema, table
+      "select column_name, * from information_schema.columns where table_schema = '{schema}' and table_name = '{table}'"
     )
   }
 
   fn preview_constraints_query(&self, schema: &str, table: &str) -> String {
     format!(
-      "select constraint_name, * from information_schema.table_constraints where table_schema = '{}' and table_name = '{}'",
-      schema, table
+      "select constraint_name, * from information_schema.table_constraints where table_schema = '{schema}' and table_name = '{table}'"
     )
   }
 
   fn preview_indexes_query(&self, schema: &str, table: &str) -> String {
-    format!("select indexname, indexdef, * from pg_indexes where schemaname = '{}' and tablename = '{}'", schema, table)
+    format!("select indexname, indexdef, * from pg_indexes where schemaname = '{schema}' and tablename = '{table}'")
   }
 
   fn preview_policies_query(&self, schema: &str, table: &str) -> String {
-    format!("select * from pg_policies where schemaname = '{}' and tablename = '{}'", schema, table)
+    format!("select * from pg_policies where schemaname = '{schema}' and tablename = '{table}'")
   }
 }
 
 impl PostgresDriver<'_> {
   pub fn new() -> Self {
-    Self { pool: None, task: None }
+    Self { pool: None, task: None, querying_conn: None, querying_pid: None }
   }
 
   fn build_connection_opts(
@@ -281,6 +342,10 @@ impl PostgresDriver<'_> {
 
 async fn query_with_pool(pool: Arc<sqlx::Pool<Postgres>>, query: String) -> Result<Rows> {
   query_with_stream(&*pool.clone(), &query).await
+}
+
+async fn query_with_conn(conn: &mut PgConnection, query: String) -> Result<Rows> {
+  query_with_stream(conn, &query).await
 }
 
 async fn query_with_stream<'a, E>(e: E, query: &'a str) -> Result<Rows>
@@ -525,7 +590,7 @@ mod tests {
   use sqlparser::{dialect::PostgreSqlDialect, parser::ParserError};
 
   use super::*;
-  use crate::database::{get_execution_type, get_first_query, ExecutionType, ParseError};
+  use crate::database::{ExecutionType, ParseError, get_execution_type, get_first_query};
 
   #[test]
   fn test_get_first_query() {
@@ -606,7 +671,7 @@ mod tests {
         (Err(ParseError::SqlParserError(msg)), Err(ParseError::SqlParserError(expected_msg))) => {
           assert_eq!(msg, expected_msg)
         },
-        _ => panic!("Unexpected result for input: {}", input),
+        _ => panic!("Unexpected result for input: {input}"),
       }
     }
   }
@@ -629,8 +694,7 @@ mod tests {
       assert_eq!(
         get_execution_type(query.to_string(), false, Driver::Postgres).unwrap().0,
         expected,
-        "Failed for query: {}",
-        query
+        "Failed for query: {query}"
       );
     }
   }
