@@ -5,7 +5,7 @@ use directories::{ProjectDirs, UserDirs};
 use lazy_static::lazy_static;
 use tracing::error;
 use tracing_error::ErrorLayer;
-use tracing_subscriber::{self, prelude::__tracing_subscriber_SubscriberExt, util::SubscriberInitExt, Layer};
+use tracing_subscriber::{self, Layer, prelude::__tracing_subscriber_SubscriberExt, util::SubscriberInitExt};
 
 const VERSION_MESSAGE: &str =
   concat!(env!("CARGO_PKG_VERSION"), "-", env!("VERGEN_GIT_DESCRIBE"), " (", env!("VERGEN_BUILD_DATE"), ")");
@@ -41,15 +41,15 @@ pub fn initialize_panic_handler() -> Result<()> {
     .into_hooks();
   eyre_hook.install()?;
   std::panic::set_hook(Box::new(move |panic_info| {
-    if let Ok(mut t) = crate::tui::Tui::new() {
-      if let Err(r) = t.exit() {
-        error!("Unable to exit Terminal: {:?}", r);
-      }
+    if let Ok(mut t) = crate::tui::Tui::new()
+      && let Err(r) = t.exit()
+    {
+      error!("Unable to exit Terminal: {:?}", r);
     }
 
     #[cfg(not(debug_assertions))]
     {
-      use human_panic::{handle_dump, print_msg, Metadata};
+      use human_panic::{Metadata, handle_dump, print_msg};
       let meta = Metadata::new(env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
         .authors(env!("CARGO_PKG_AUTHORS").replace(':', ", "))
         .homepage(env!("CARGO_PKG_HOMEPAGE"));
@@ -78,40 +78,37 @@ pub fn initialize_panic_handler() -> Result<()> {
 }
 
 pub fn get_favorites_dir() -> PathBuf {
-  let directory = if let Some(s) = FAVORITES_FOLDER.clone() {
+  if let Some(s) = FAVORITES_FOLDER.clone() {
     s
   } else if let Some(proj_dirs) = project_directory() {
     proj_dirs.data_local_dir().to_path_buf()
   } else {
     PathBuf::from(".").join(".favorites")
-  };
-  directory
+  }
 }
 
 pub fn get_data_dir() -> PathBuf {
-  let directory = if let Some(s) = DATA_FOLDER.clone() {
+  if let Some(s) = DATA_FOLDER.clone() {
     s
   } else if let Some(proj_dirs) = project_directory() {
     proj_dirs.data_local_dir().to_path_buf()
   } else {
     PathBuf::from(".").join(".data")
-  };
-  directory
+  }
 }
 
 pub fn get_config_dir() -> PathBuf {
-  let directory = if let Some(s) = CONFIG_FOLDER.clone() {
+  if let Some(s) = CONFIG_FOLDER.clone() {
     s
   } else if let Some(proj_dirs) = project_directory() {
     proj_dirs.config_local_dir().to_path_buf()
   } else {
     PathBuf::from(".").join(".config")
-  };
-  directory
+  }
 }
 
 pub fn get_export_dir() -> PathBuf {
-  let directory = if let Some(s) = EXPORT_FOLDER.clone() {
+  if let Some(s) = EXPORT_FOLDER.clone() {
     s
   } else if let Some(user_dir) = user_directory() {
     if let Some(download_dir) = user_dir.download_dir() {
@@ -121,8 +118,7 @@ pub fn get_export_dir() -> PathBuf {
     }
   } else {
     PathBuf::from(".").join(".export")
-  };
-  directory
+  }
 }
 
 pub fn initialize_logging() -> Result<()> {
@@ -130,12 +126,15 @@ pub fn initialize_logging() -> Result<()> {
   std::fs::create_dir_all(directory.clone())?;
   let log_path = directory.join(LOG_FILE.clone());
   let log_file = std::fs::File::create(log_path)?;
-  std::env::set_var(
-    "RUST_LOG",
-    std::env::var("RUST_LOG")
-      .or_else(|_| std::env::var(LOG_ENV.clone()))
-      .unwrap_or_else(|_| format!("{}=info", env!("CARGO_CRATE_NAME"))),
-  );
+  // TODO: Audit that the environment access only happens in single-threaded code.
+  unsafe {
+    std::env::set_var(
+      "RUST_LOG",
+      std::env::var("RUST_LOG")
+        .or_else(|_| std::env::var(LOG_ENV.clone()))
+        .unwrap_or_else(|_| format!("{}=info", env!("CARGO_CRATE_NAME"))),
+    )
+  };
   let file_subscriber = tracing_subscriber::fmt::layer()
     .with_file(true)
     .with_line_number(true)
@@ -154,7 +153,7 @@ pub fn initialize_logging() -> Result<()> {
 /// this can be customized.
 #[macro_export]
 macro_rules! trace_dbg {
-  (target: $target:expr, level: $level:expr, $ex:expr) => {{
+  (target: $target:expr_2021, level: $level:expr_2021, $ex:expr_2021) => {{
     match $ex {
       value => {
         tracing::event!(target: $target, $level, ?value, stringify!($ex));
@@ -162,13 +161,13 @@ macro_rules! trace_dbg {
       },
     }
   }};
-  (level: $level:expr, $ex:expr) => {
+  (level: $level:expr_2021, $ex:expr_2021) => {
     trace_dbg!(target: module_path!(), level: $level, $ex)
   };
-  (target: $target:expr, $ex:expr) => {
+  (target: $target:expr_2021, $ex:expr_2021) => {
     trace_dbg!(target: $target, level: tracing::Level::DEBUG, $ex)
   };
-  ($ex:expr) => {
+  ($ex:expr_2021) => {
     trace_dbg!(level: tracing::Level::DEBUG, $ex)
   };
 }
